+++ packages/react/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.0]
+
+### Deprecations
+
+The following deprecated API has been removed in this release:
+
+- The `PageHeading` component. Use a `Heading` instead.
+- The `color` prop of `Menu.Link`.
+- The `size` value `level-6` of `Heading`. Use `level-5` instead.
+- The `size` values `heading-0` and `heading-6` of `Icon`. Use `heading-1` and `heading-5` instead.
+- The `headingLevel` value `1` of `Accordion`. Use `2`, `3`, or `4` instead.
+- The `visuallyHiddenLabel`, `visuallyHiddenLabelId`, `nextVisuallyHiddenLabel`, and `previousVisuallyHiddenLabel` props of `Pagination`. Use `accessibleName`, `accessibleNameId`, `nextAccessibleName`, and `previousAccessibleName` instead.
+
 ## [3.3.0](https://github.com/Amsterdam/design-system/compare/design-system-react-v3.2.1...design-system-react-v3.3.0) (2026-04-03)
 
 
@@ -92,7 +105,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Deprecations
 
-The following prop will be removed in release 4.0.0:
+The following prop has been removed in release 4.0.0:
 
 * **Menu.Link**: `color`
 
@@ -133,7 +146,7 @@ The following prop will be removed in release 4.0.0:
 
 ### Deprecations
 
-The following props and prop values will be removed in release 4.0.0:
+The following props and prop values have been removed in release 4.0.0:
 
 * **Accordion**: `headingLevel` 1
 * **Heading**: `size` 'level-6'